requirements.txt
pyyaml >= 3.13
Pillow >= 2.1.0


